:factory: move protos service to pkg folder so that it can be imported by API
