+++ pricing-service/internal/server/rates.go
@@ -4,7 +4,7 @@ import (
 	context "context"
 
 	"github.com/sirupsen/logrus"
-	protos "github.com/zenichi/green-api/pricing-service/internal/protos/rates"
+	protos "github.com/zenichi/green-api/pricing-service/pkg/protos/rates"
 )
 
 // Rates is a grpc server, it implements methods defined by RateServiceServer
