+++ pricing-service/pkg/protos/rates.proto
